% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ascifit.R
\name{erf}
\alias{erf}
\title{The Error Function (or Gaussian Error function).}
\usage{
erf(x)
}
\arguments{
\item{x}{: A non-negative real number.}
}
\value{
: A non-negative real value corresponding to the error function.
}
\description{
For any \eqn{x \geq 0}, this returns \eqn{\operatorname{Erf}(x)}. Here the
function \eqn{\operatorname{Erf}(x)} is defined to be
\deqn{\operatorname{Erf}(x) := \frac{2}{\sqrt{\pi}} \int_{0}^{x} e^{-t^{2}} d t}
}
\examples{
\dontrun{
erf(0) # should return 0
}
}
