% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ascifit.R
\name{mean_theta_sigma}
\alias{mean_theta_sigma}
\title{The Inverse Folded Normal mean value, for a fixed standard deviation.}
\usage{
mean_theta_sigma(theta, sigma)
}
\arguments{
\item{theta}{: A non-negative real number.}

\item{sigma}{: A positive real number.}
}
\value{

}
\description{
The Inverse Folded Normal mean value, for a fixed standard deviation.
}
